refactor: data type for BrighnessInterval and usage of Constants

+++ DigitalFootprintExplorer/shared/src/commonMain/kotlin/ch/zhaw/init/digitalfootprintexplorer/digitalfootprintexplorer/model/input/DisplayInput.kt
@@ -1,8 +1,8 @@
 package ch.zhaw.init.digitalfootprintexplorer.digitalfootprintexplorer.model.input
 
 data class BrightnessInterval(
-    val normalizedBrightness: Float,
-    val durationH: Float
+    val normalizedBrightness: Double,
+    val durationH: Double
 )
 
 data class DisplayInput(
